feat(mission) replace int distance to double

+++ src/main/java/artemis/model/mission/Mission.java
@@ -4,11 +4,11 @@ public abstract class Mission {
 
     private String name;
     private boolean manned;
-    private double distance;
+    private long distance;
     private String time;
     private double fuelCoefficient;
 
-    public Mission(String name, boolean manned, double distance, String time, double fuelCoefficient){
+    public Mission(String name, boolean manned, long distance, String time, double fuelCoefficient){
         this.name = name;
         this.manned = manned;
         this.distance = distance;
@@ -20,7 +20,7 @@ public abstract class Mission {
 
     public String getName(){return name;}
     public boolean getManned(){return manned;}
-    public double getDistance(){return distance;}
+    public long getDistance(){return distance;}
     public String getTime(){return time;}
     public double getFuelCoefficient(){return fuelCoefficient;}
